Load FTN Charting Data

FTN Data manually charts plays and has graciously provided a subset of their charting data to be published via the nflverse. Data is available from the 2022 season onwards and is charted within 48 hours following each game. This data is released under the CC-BY-SA 4.0 Creative Commons license and attribution must be made to FTN Data via nflverse

Usage

load_ftn_charting(
  seasons = most_recent_season(),
  file_type = getOption("nflreadr.prefer", default = "rds")
)

Source

FTNData.com

Arguments

seasons

a numeric vector of seasons to return, defaults to most recent season. If set to TRUE, returns all available data. Data available from 2022 onwards.

file_type

One of c("rds", "qs", "csv", "parquet"). Can also be set globally with options(nflreadr.prefer)

Value

Play-level manual charting data from FTN Data

See also

Author

FTN Data

Examples

# \donttest{
  try({ # prevents cran errors
  load_ftn_charting()
  })
#> ── nflverse FTN Charting Data - please attribute to 'ftndata.com via nflverse' ─
#> ℹ Data updated: 2023-12-21 18:11:43 UTC
#> # A tibble: 38,089 × 29
#>    ftn_game_id nflverse_game_id season  week ftn_play_id nflverse_play_id
#>  *       <int> <chr>             <int> <int>       <int>            <int>
#>  1        6164 2023_01_DET_KC     2023     1     1009770               40
#>  2        6164 2023_01_DET_KC     2023     1     1009771               56
#>  3        6164 2023_01_DET_KC     2023     1     1009772               78
#>  4        6164 2023_01_DET_KC     2023     1     1009773              101
#>  5        6164 2023_01_DET_KC     2023     1     1009774              124
#>  6        6164 2023_01_DET_KC     2023     1     1009775              150
#>  7        6164 2023_01_DET_KC     2023     1     1009776              172
#>  8        6164 2023_01_DET_KC     2023     1     1009777              194
#>  9        6164 2023_01_DET_KC     2023     1     1009778              219
#> 10        6164 2023_01_DET_KC     2023     1     1009779              244
#> # ℹ 38,079 more rows
#> # ℹ 23 more variables: starting_hash <chr>, qb_location <chr>,
#> #   n_offense_backfield <int>, n_defense_box <int>, is_no_huddle <lgl>,
#> #   is_motion <lgl>, is_play_action <lgl>, is_screen_pass <lgl>, is_rpo <lgl>,
#> #   is_trick_play <lgl>, is_qb_out_of_pocket <lgl>,
#> #   is_interception_worthy <lgl>, is_throw_away <lgl>, read_thrown <chr>,
#> #   is_catchable_ball <lgl>, is_contested_ball <lgl>, …
# }