Load Advanced Passing Stats from PFR

Loads player level season stats provided by Pro Football Reference starting with the 2019 season.

load_pfr_passing(seasons = TRUE)

Arguments

seasons
a numeric vector specifying what seasons to return, if TRUE returns all available data

Value

A tibble of season-level player statistics provided by Pro Football Reference.

See also

Examples

# \donttest{
  load_pfr_passing()

#> # A tibble: 214 × 28
#>    player      team  pass_attempts batted_balls throwaways spikes drops drop_pct
#>    <chr>       <chr>         <dbl>        <dbl>      <dbl>  <dbl> <dbl>    <dbl>
#>  1 Jameis Win… TB              626           11         22      2    23      3.8
#>  2 Dak Presco… DAL             596            9         14      1    36      6.2
#>  3 Jared Goff  LAR             626           17         15      1    27      4.4
#>  4 Philip Riv… LAC             591           13         16      3    25      4.4
#>  5 Matt Ryan   ATL             616           19         27      1    17      2.9
#>  6 Russell Wi… SEA             516            9         20      3    17      3.4
#>  7 Tom Brady   NE              613            6         40      0    34      5.9
#>  8 Derek Carr  OAK             513            9         28      2    25      5.2
#>  9 Carson Wen… PHI             607           16         10      1    28      4.7
#> 10 Patrick Ma… KC              484            6         19      0    25      5.4
#> # … with 204 more rows, and 20 more variables: bad_throws <dbl>,
#> #   bad_throw_pct <dbl>, on_tgt_throws <dbl>, on_tgt_pct <dbl>, season <int>,
#> #   pfr_id <chr>, pocket_time <dbl>, times_blitzed <dbl>, times_hurried <dbl>,
#> #   times_hit <dbl>, times_pressured <dbl>, pressure_pct <dbl>,
#> #   rpo_plays <dbl>, rpo_yards <dbl>, rpo_pass_att <dbl>, rpo_pass_yards <dbl>,
#> #   rpo_rush_att <dbl>, rpo_rush_yards <dbl>, pa_pass_att <dbl>,
#> #   pa_pass_yards <dbl>

# }