Load Snap Counts from PFR

Loads game level snap counts stats provided by Pro Football Reference starting with the 2012 season.

Usage

load_snap_counts(
  seasons = most_recent_season(),
  file_type = getOption("nflreadr.prefer", default = "rds")
)

Arguments

seasons

a numeric vector specifying what seasons to return, if TRUE returns all available data

file_type

One of c("rds", "qs", "csv", "parquet"). Can also be set globally with options(nflreadr.prefer)

Value

A tibble of game-level snap counts provided by Pro Football Reference.

See also

Examples

# \donttest{
  try({ # prevents CRAN errors
  load_snap_counts()
  })
#> ── nflverse snap counts ────────────────────────────────────────────────────────
#> ℹ Data updated: 2025-02-10 18:32:34 UTC
#> # A tibble: 26,600 × 16
#>    game_id      pfr_game_id season game_type  week player pfr_player_id position
#>    <chr>        <chr>        <int> <chr>     <int> <chr>  <chr>         <chr>   
#>  1 2024_01_ARI… 202409080b…   2024 REG           1 Spenc… BrowSp00      T       
#>  2 2024_01_ARI… 202409080b…   2024 REG           1 O'Cyr… TorrOC00      G       
#>  3 2024_01_ARI… 202409080b…   2024 REG           1 Dion … DawkDi00      T       
#>  4 2024_01_ARI… 202409080b…   2024 REG           1 David… EdwaDa01      G       
#>  5 2024_01_ARI… 202409080b…   2024 REG           1 Josh … AlleJo02      QB      
#>  6 2024_01_ARI… 202409080b…   2024 REG           1 Conno… McGoCo01      C       
#>  7 2024_01_ARI… 202409080b…   2024 REG           1 Dalto… KincDa00      TE      
#>  8 2024_01_ARI… 202409080b…   2024 REG           1 Keon … ColeKe02      WR      
#>  9 2024_01_ARI… 202409080b…   2024 REG           1 James… CookJa01      RB      
#> 10 2024_01_ARI… 202409080b…   2024 REG           1 Mack … HollMa00      WR      
#> # ℹ 26,590 more rows
#> # ℹ 8 more variables: team <chr>, opponent <chr>, offense_snaps <dbl>,
#> #   offense_pct <dbl>, defense_snaps <dbl>, defense_pct <dbl>, st_snaps <dbl>,
#> #   st_pct <dbl>
# }