List all available nflverse releases

This functions lists all nflverse data releases that are available in the nflverse-data repo. Release names can be used for downloads in nflverse_download().

Usage

nflverse_releases(.token = "default")

Arguments

.token

a GitHub API token, "default" uses gh::gh_token()

Value

A dataframe containing release names, release descriptions, and other relevant release information.

Examples

# \donttest{

try( # avoids cran failures, can skip in normal usage
nflverse_releases()
)
#> ── nflverse release listing ────────────────────────────────────────────────────
#> ℹ Data updated: 2025-06-29 20:11:25 UTC
#> # A tibble: 20 × 8
#>    release_name   release_description timestamp   rds parquet   csv csv_gz   zip
#>    <chr>          <chr>               <chr>     <int>   <int> <int>  <int> <int>
#>  1 players_compo… "Component files f… 2025-06-…     2       2     2      2     0
#>  2 contracts      "OverTheCap contra… 2025-06-…     2       1     0      1     0
#>  3 rosters        "Roster data, acce… 2025-06-…   106     106   106      2     0
#>  4 weekly_rosters "Week-level roster… 2025-06-…    24      24    24      2     0
#>  5 players        "Player informatio… 2025-05-…     1       1     1      1     0
#>  6 player_stats   "Play by play data… 2025-05-…   365     367   365    366     0
#>  7 draft_picks    "Draft picks datin… 2025-05-…     1       1     1      1     0
#>  8 pfr_advstats   "PFR Adv Stats dat… 2025-04-…    60      32    32      8     0
#>  9 pbp            "Play by play data… 2025-04-…    26      26    26     26    23
#> 10 misc           "Various bits of d… 2025-04-…     7       1     6      0     0
#> 11 combine        "NFL Combine Data … 2025-03-…     1       1     1      1     0
#> 12 espn_data      "ESPN Stats"        2025-02-…     2       2     2      2     0
#> 13 nextgen_stats  "NFL Next Gen Stat… 2025-02-…    30       3     0     30     0
#> 14 ftn_charting   "Charting Data pro… 2025-02-…     3       3     3      0     0
#> 15 depth_charts   "Depth chart data,… 2025-02-…    24      24    24      1     0
#> 16 injuries       "Injuries data, ac… 2025-02-…    16      16    16      2     0
#> 17 snap_counts    "Snap counts data,… 2025-02-…    13      13    14      1     0
#> 18 test           "for internal test… 2025-01-…     4       1     1      0     0
#> 19 officials      "Data about which … 2025-01-…     1       1     1      1     0
#> 20 pbp_participa… "Participation dat… 2024-02-…     8       8     8      0     0
# }