Load FTN Charting Data

FTN Data manually charts plays and has graciously provided a subset of their charting data to be published via the nflverse. Data is available from the 2022 season onwards and is charted within 48 hours following each game. This data is released under the CC-BY-SA 4.0 Creative Commons license and attribution must be made to FTN Data via nflverse

Usage

load_ftn_charting(
  seasons = most_recent_season(),
  file_type = getOption("nflreadr.prefer", default = "rds")
)

Source

FTNData.com

Arguments

seasons

a numeric vector of seasons to return, defaults to most recent season. If set to TRUE, returns all available data. Data available from 2022 onwards.

file_type

One of c("rds", "qs", "csv", "parquet"). Can also be set globally with options(nflreadr.prefer)

Value

Play-level manual charting data from FTN Data

See also

Author

FTN Data

Examples

# \donttest{
  try({ # prevents cran errors
  load_ftn_charting()
  })
#> ── nflverse FTN Charting Data - please attribute to 'ftndata.com via nflverse' ─
#> ℹ Data updated: 2025-02-14 18:20:48 UTC
#> # A tibble: 48,031 × 29
#>    ftn_game_id nflverse_game_id season  week ftn_play_id nflverse_play_id
#>  *       <int> <chr>             <int> <int>       <int>            <int>
#>  1        6449 2024_01_BAL_KC     2024     1     1057995               43
#>  2        6449 2024_01_BAL_KC     2024     1     1057996               60
#>  3        6449 2024_01_BAL_KC     2024     1     1057997               82
#>  4        6449 2024_01_BAL_KC     2024     1     1057998              118
#>  5        6449 2024_01_BAL_KC     2024     1     1057999              148
#>  6        6449 2024_01_BAL_KC     2024     1     1058000              183
#>  7        6449 2024_01_BAL_KC     2024     1     1058001              205
#>  8        6449 2024_01_BAL_KC     2024     1     1058002              239
#>  9        6449 2024_01_BAL_KC     2024     1     1058003              273
#> 10        6449 2024_01_BAL_KC     2024     1     1058004              298
#> # ℹ 48,021 more rows
#> # ℹ 23 more variables: starting_hash <chr>, qb_location <chr>,
#> #   n_offense_backfield <int>, n_defense_box <int>, is_no_huddle <lgl>,
#> #   is_motion <lgl>, is_play_action <lgl>, is_screen_pass <lgl>, is_rpo <lgl>,
#> #   is_trick_play <lgl>, is_qb_out_of_pocket <lgl>,
#> #   is_interception_worthy <lgl>, is_throw_away <lgl>, read_thrown <chr>,
#> #   is_catchable_ball <lgl>, is_contested_ball <lgl>, …
# }