List all available nflverse releases

This functions lists all nflverse data releases that are available in the nflverse-data repo. Release names can be used for downloads in nflverse_download().

Usage

nflverse_releases(.token = "default")

Arguments

.token

a GitHub API token, "default" uses gh::gh_token()

Value

A dataframe containing release names, release descriptions, and other relevant release information.

Examples

# \dontshow{
.for_cran()
# Change option for better output
old <- options(piggyback.verbose = FALSE)
# }
# \donttest{

try( # avoids cran failures, can skip in normal usage
nflverse_releases()
)
#> ── nflverse release listing ────────────────────────────────────────────────────
#> ℹ Data updated: 2023-09-19 02:38:38 UTC
#> # A tibble: 19 × 8
#>    release_name   release_description timestamp   rds parquet   csv csv_gz   zip
#>    <chr>          <chr>               <chr>     <int>   <int> <int>  <int> <int>
#>  1 ftn_charting   "Charting Data pro… 2023-09-…     2       2     2      0     0
#>  2 snap_counts    "Snap counts data,… 2023-09-…    12      12    13      0     0
#>  3 pbp_participa… "Participation dat… 2023-09-…     8       8     8      0     0
#>  4 player_stats   "Play by play data… 2023-09-…    78      78    78     75     0
#>  5 pbp            "Play by play data… 2023-09-…    25      25    25     25    23
#>  6 nextgen_stats  "NFL Next Gen Stat… 2023-09-…    27       3     0     27     0
#>  7 pfr_advstats   "PFR Adv Stats dat… 2023-09-…    52      28    28      0     0
#>  8 contracts      "OverTheCap contra… 2023-09-…     2       1     0      1     0
#>  9 players_compo… "Component files f… 2023-09-…     2       2     2      0     0
#> 10 injuries       "Injuries data, ac… 2023-09-…    15      15    15      0     0
#> 11 depth_charts   "Depth chart data,… 2023-09-…    23      23    23      0     0
#> 12 rosters        "Roster data, acce… 2023-09-…   104     104   104      0     0
#> 13 weekly_rosters "Week-level roster… 2023-09-…    22      22    22      0     0
#> 14 officials      "Data about which … 2023-09-…     1       1     1      0     0
#> 15 misc           "Various bits of d… 2023-09-…     7       1     6      0     0
#> 16 draft_picks    "Draft picks datin… 2023-09-…     1       1     1      0     0
#> 17 players        "Player informatio… 2023-09-…     1       1     1      0     0
#> 18 combine        "NFL Combine Data … 2023-05-…     1       1     1      0     0
#> 19 test           "for internal test… 2022-05-…     1       0     1      0     0
# }
# \dontshow{
# Restore old options
options(old)
# }