Get a Situation Report on System, nflverse/ffverse Package Versions and Dependencies

This function gives a quick overview of the versions of R and the operating system as well as the versions of nflverse/ffverse packages and their dependencies. It's primarily designed to help you get a quick idea of what's going on when you're helping someone else debug a problem.

nflverse_sitrep(
  pkg = c("nflreadr", "nflfastR", "nflseedR", "nfl4th", "nflplotR"),
  recursive = TRUE
)

ffverse_sitrep(pkg = c("ffscrapr", "ffsimulator", "ffpros"), recursive = TRUE)

.sitrep(pkg, recursive = TRUE, header = "")

Arguments

pkg
a character vector naming installed packages, or NULL (the default) meaning all nflverse packages. The function checks internally if all packages are installed and informs if that is not the case.

recursive
a logical indicating whether dependencies of pkg and their dependencies (and so on) should be included. Can also be a character vector listing the types of dependencies, a subset of c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances"). Character string "all" is shorthand for that vector, character string "most" for the same vector without "Enhances", character string "strong" (default) for the first three elements of that vector.

header
a string that is printed in the horizontal separation lines and used to differentiate between nflverse and ffverse output.

Examples

# \donttest{
try({
nflverse_sitrep()
ffverse_sitrep()
.sitrep("cachem")
})

#> ℹ You've asked for the packages nflfastR, nflseedR, nfl4th, and nflplotR which are not installed. 
#> They are skipped.

#> ── System Info ─────────────────────────────────────────────────────────────────
#> • R version 4.1.2 (2021-11-01)   • Running under: Ubuntu 20.04.3 LTS
#> ── nflverse Packages ───────────────────────────────────────────────────────────
#> • nflreadr (1.1.1.02)
#> ── nflverse Dependencies ───────────────────────────────────────────────────────
#> • cachem     (1.0.6)   • memoise       (2.0.0)   • Rcpp         (1.0.7)   
#> • curl       (4.3.2)   • qs            (0.25.1)  • RcppParallel (5.1.4)   
#> • data.table (1.14.2)  • RApiSerialize (0.1.0)   • rlang        (0.4.12)  
#> • fastmap    (1.1.0)   • rappdirs      (0.3.3)   • stringfish   (0.15.4)  

#> ℹ You've asked for the packages ffscrapr, ffsimulator, and ffpros which are not installed. 
#> They are skipped.

#> Error in utils::sessionInfo(packages) : no valid packages were specified

# }